chore: refactor analyser code
