+++ src/components/analyser-pane/analyser-pane.css
@@ -1,8 +1,36 @@
+@import "../../css/colors.css";
 @import "../../css/units.css";
+@import "../../css/z-index.css";
 
 .analyser-pane {
-    /* Makes columns for the sprite library selector + and the stage selector */
-    display: flex;
-    flex-direction: row;
     flex-grow: 1;
+    position: relative;
+    font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
+    background-color: $ui-secondary;
+    border-top-right-radius: $space;
+    border-top-left-radius: $space;
+    border-color: $ui-black-transparent;
+    border-width: 1px;
+    border-style: solid;
+    border-bottom: 0;
+}
+
+.chart-wrapper {
+    box-sizing: border-box;
+    width: 100%;
+    margin: calc($space / 2);
+}
+
+.items-wrapper {
+    display: flex;
+    flex-wrap: wrap;
+    flex-direction: row;
+
+    padding-top: calc($space / 2);
+    padding-left: calc($space / 2);
+    padding-right: calc($space / 2);
+    padding-bottom: $space;
+
+    overflow-x: hidden;
+    overflow-y: auto;
 }
