+++ src/components/chart-info/chart-info.css
