+++ src/components/chart-info/chart-info.jsx
