+++ src/components/chart-selector-item/chart-selector-item.css
