+++ src/containers/gui.jsx
@@ -115,7 +115,9 @@ GUI.propTypes = {
     isShowingProject: PropTypes.bool,
     loadingStateVisible: PropTypes.bool,
     onProjectLoaded: PropTypes.func,
+    onCloseAlert: PropTypes.func,
     onSeeCommunity: PropTypes.func,
+    onShowAlert: PropTypes.func,
     onStorageInit: PropTypes.func,
     onUpdateProjectId: PropTypes.func,
     onVmInit: PropTypes.func,
